# Optimistic route registry locks