--- a/registry/registry_benchmark_test.go
+++ b/registry/registry_benchmark_test.go
@@ -1,7 +1,9 @@
 package registry_test
 
 import (
+	"context"
 	"fmt"
+	"sync"
 	"testing"
 	"time"
 
@@ -37,7 +39,7 @@ func setupLogger() logger.Logger {
 	l := logger.NewLogger(
 		"test",
 		"unix-epoch",
-		zap.InfoLevel,
+		zap.WarnLevel,
 		zap.Output(zap.MultiWriteSyncer(sink, zap.AddSync(ginkgo.GinkgoWriter))),
 		zap.ErrorOutput(zap.MultiWriteSyncer(sink, zap.AddSync(ginkgo.GinkgoWriter))),
 	)
@@ -82,5 +84,58 @@ func BenchmarkRegisterWithOneRoute(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		r.Register("foo.example.com", fooEndpoint)
 	}
+	b.ReportAllocs()
+}
+
+func BenchmarkRegisterWithConcurrentLookupWith100kRoutes(b *testing.B) {
+	r := registry.NewRouteRegistry(testLogger, configObj, reporter)
+	maxRoutes := 100000
+	routeUris := make([]route.Uri, maxRoutes)
+
+	for i := 0; i < maxRoutes; i++ {
+		routeUris[i] = route.Uri(fmt.Sprintf("foo%d.example.com", i))
+		r.Register(routeUris[i], fooEndpoint)
+	}
+
+	lookupCounts := make(chan uint)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	numLookupers := 10
+	var wg sync.WaitGroup
+	for i := 0; i < numLookupers; i++ {
+		wg.Add(1)
+		go func(ctx context.Context, wg *sync.WaitGroup) {
+			var lookups uint
+			for i := 0; ; i++ {
+				select {
+				case <-ctx.Done():
+					wg.Done()
+					lookupCounts <- lookups
+					return
+				default:
+					routeUri := routeUris[i%maxRoutes]
+					r.Lookup(routeUri)
+					lookups++
+				}
+			}
+		}(ctx, &wg)
+	}
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		r.Register(routeUris[i%maxRoutes], fooEndpoint)
+	}
+	b.Elapsed()
+
+	cancel()
+	wg.Wait()
+
+	var lookupCount uint
+	for i := 0; i < numLookupers; i++ {
+		lookupCount += <-lookupCounts
+	}
+
+	b.Logf("Looked up %d routes concurrently, registered %d", lookupCount, b.N)
 	b.ReportAllocs()
 }