@peanball 2023-12-12
@domdom82, it already runs with the race detector and had found some races in earlier development that are now fixed.

I was very happy to see the race detector in action :)